add opt-in to enable support for passing multiple collection infos
Intended to avoid a driver incorrectly using the old protocol for `MONGOCRYPT_CTX_NEED_MONGO_COLLINFO(_WITH_DB)`. The old protocol only requires passing one collinfo. libmongocrypt interprets "no schema fed" as "no schema is configured on the server".

The opt-in is intended for drivers to signal the new protocol is used. If not opted-in, commands referencing multiple collections result in error.